增加mlp实验

+++ data_prepare/time_series.py
@@ -10,7 +10,7 @@ from config import DATETIME_COLUMN_NAMES, TIMESTAMP_COLUMN_NAMES
 class TimeSeries(object):
     def __init__(self, data: pd.DataFrame):
         super(TimeSeries, self).__init__()
-        self.data = data.copy(deep=True)
+        self.data = data.astype(float).copy(deep=True)
         # 判断是否有时间索引
         if not isinstance(self.data.index, pd.DatetimeIndex):
             for ts_column in DATETIME_COLUMN_NAMES:
